// src/modules/adm/empresa/dto/update-empresa.dto.ts
export class UpdateEmpresaDto {
  vch_nomemp?:        string;
  vch_cgcemp?:        string;
  vch_insestemp?:     string;
  int_codstaemp?:     number;
  int_codati?:        number;
  int_codcid?:        number;
  chr_codmtzfilemp?:  string;
  chr_vldbcoemp?:     string;
  vch_nomfntemp?:     string;
  chr_cepemp?:        string;
  chr_estemp?:        string;
  vch_cidemp?:        string;
  vch_baiemp?:        string;
  vch_endemp?:        string;
  int_numendemp?:     number;
  vch_cplendemp?:     string;
  vch_tlfemp?:        string;
  vch_faxemp?:        string;
  vch_obsemp?:        string;
  vch_nomctoemp?:     string;
}